# api/openapi.yaml
openapi: 3.1.0
info:
  title: 神州 HR 考勤系统 API
  version: 1.9.0-wave3
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    V1.9 WAVE-3 契约：在完整保留 WAVE-1 本地账号、登录会话、权限、审计与
    版本化规则配置底座，以及 WAVE-2 组织、员工期初导入与本地人员权威的基础上，
    增加考勤地点、考勤组、员工期限分配、不可变班次版本、工作日历和基础策略绑定。
    所有 ID 均按 string 传输；响应和示例不包含密码、密码哈希、session token 或 reset token。
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /policy-templates:
    get:
      tags: [Policy]
      operationId: listPolicyTemplates
      summary: 分页查询策略模板
      x-capability: POLICY:READ
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
        - name: query
          in: query
          schema: { type: string, maxLength: 100 }
        - name: status
          in: query
          schema: { type: string, enum: [ACTIVE, INACTIVE] }
      responses:
        '200':
          description: 策略模板
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyTemplatePage'
        '403':
          $ref: '#/components/responses/AccessDenied'
    post:
      tags: [Policy]
      operationId: createPolicyTemplate
      summary: 创建受控字段策略模板
      x-capability: POLICY:CREATE
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyTemplateCreateRequest'
      responses:
        '201':
          description: 模板已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyTemplateDetail'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '409':
          $ref: '#/components/responses/Conflict'
  /policy-templates/{templateId}:
    get:
      tags: [Policy]
      operationId: getPolicyTemplate
      summary: 获取策略模板详情
      x-capability: POLICY:READ
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '200':
          description: 模板详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyTemplateDetail'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
  /policy-templates/{templateId}/versions:
    get:
      tags: [Policy]
      operationId: listPolicyVersions
      summary: 查询模板版本时间线
      x-capability: POLICY:READ
      parameters:
        - $ref: '#/components/parameters/TemplateId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 版本列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyVersionPage'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
    post:
      tags: [Policy]
      operationId: createPolicyDraft
      summary: 创建新的策略草稿版本
      x-capability: POLICY:CREATE
      parameters:
        - $ref: '#/components/parameters/TemplateId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyDraftCreateRequest'
      responses:
        '201':
          description: 草稿已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyVersionDetail'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '409':
          $ref: '#/components/responses/Conflict'
  /policy-templates/{templateId}/versions/{versionId}:
    get:
      tags: [Policy]
      operationId: getPolicyVersion
      summary: 获取策略版本详情
      x-capability: POLICY:READ
      parameters:
        - $ref: '#/components/parameters/TemplateId'
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200':
          description: 版本详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyVersionDetail'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
    patch:
      tags: [Policy]
      operationId: updatePolicyDraft
      summary: 修改草稿中的受控字段
      x-capability: POLICY:EDIT
      parameters:
        - $ref: '#/components/parameters/TemplateId'
        - $ref: '#/components/parameters/VersionId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyDraftUpdateRequest'
      responses:
        '200':
          description: 更新后的草稿
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyVersionDetail'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '409':
          $ref: '#/components/responses/StaleVersion'
  /policy-templates/{templateId}/versions/{versionId}/scope-bindings:
    put:
      tags: [Policy]
      operationId: replacePolicyScopeBindings
      summary: 更新草稿的公司、地点、考勤组或政策组作用范围
      x-capability: POLICY:EDIT
      parameters:
        - $ref: '#/components/parameters/TemplateId'
        - $ref: '#/components/parameters/VersionId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScopeBindingsUpdateRequest'
      responses:
        '200':
          description: 更新后的版本
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyVersionDetail'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '409':
          $ref: '#/components/responses/StaleVersion'
  /policy-templates/{templateId}/versions/{versionId}/validate:
    post:
      tags: [Policy]
      operationId: validatePolicyVersion
      summary: 校验受控字段、枚举、版本和有效期
      x-capability: POLICY:VALIDATE
      parameters:
        - $ref: '#/components/parameters/TemplateId'
        - $ref: '#/components/parameters/VersionId'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: 校验结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyValidationResult'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '409':
          $ref: '#/components/responses/StaleVersion'
  /policy-templates/{templateId}/versions/{versionId}/conflicts:
    post:
      tags: [Policy]
      operationId: checkPolicyConflicts
      summary: 检查作用范围、生效期和优先级冲突
      x-capability: POLICY:VALIDATE
      parameters:
        - $ref: '#/components/parameters/TemplateId'
        - $ref: '#/components/parameters/VersionId'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: 冲突结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyConflictResult'
        '403':
          $ref: '#/components/responses/AccessDenied'
  /policy-templates/{templateId}/versions/{versionId}/simulate:
    post:
      tags: [Policy]
      operationId: simulatePolicyVersion
      summary: 对受控合成输入执行样例试算
      x-capability: POLICY:SIMULATE
      parameters:
        - $ref: '#/components/parameters/TemplateId'
        - $ref: '#/components/parameters/VersionId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicySimulationRequest'
      responses:
        '200':
          description: 试算结果；不执行任意脚本
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicySimulationResult'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/AccessDenied'
  /policy-templates/{templateId}/versions/{versionId}/impact-preview:
    post:
      tags: [Policy]
      operationId: previewPolicyImpact
      summary: 发布前预览范围、期间、冻结保护和影响摘要
      x-capability: POLICY:SIMULATE
      parameters:
        - $ref: '#/components/parameters/TemplateId'
        - $ref: '#/components/parameters/VersionId'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: 影响预览
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyImpactPreview'
        '403':
          $ref: '#/components/responses/AccessDenied'
  /policy-templates/{templateId}/versions/{versionId}/publish:
    post:
      tags: [Policy]
      operationId: publishPolicyVersion
      summary: 发布通过校验且无冲突的策略版本
      x-capability: POLICY:PUBLISH
      parameters:
        - $ref: '#/components/parameters/TemplateId'
        - $ref: '#/components/parameters/VersionId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishPolicyRequest'
      responses:
        '200':
          description: 已发布版本
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyVersionDetail'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '409':
          $ref: '#/components/responses/Conflict'
  /policy-templates/{templateId}/versions/{versionId}/deactivate:
    post:
      tags: [Policy]
      operationId: deactivatePolicyVersion
      summary: 追加记录并停用策略版本
      x-capability: POLICY:DEACTIVATE
      parameters:
        - $ref: '#/components/parameters/TemplateId'
        - $ref: '#/components/parameters/VersionId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '200':
          description: 已停用版本
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyVersionDetail'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '409':
          $ref: '#/components/responses/Conflict'
  /policy-templates/{templateId}/versions/{versionId}/rollback:
    post:
      tags: [Policy]
      operationId: rollbackPolicyVersion
      summary: 从历史快照创建并发布新的回滚版本
      x-capability: POLICY:ROLLBACK
      parameters:
        - $ref: '#/components/parameters/TemplateId'
        - $ref: '#/components/parameters/VersionId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RollbackPolicyRequest'
      responses:
        '201':
          description: 新的回滚版本
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyVersionDetail'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SHENZHOUHR_SESSION
      description: 本系统本地账号安全会话；开发身份头不属于正式认证路径。
...
  parameters:
    CsrfToken:
      name: X-CSRF-TOKEN
      in: header
      required: true
      schema: { type: string, minLength: 32, maxLength: 256 }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Sort:
      name: sort
      in: query
      schema:
        type: string
        enum: [createdAt,updatedAt,username,status,occurredAt,templateCode,versionNumber]
...
    TemplateId:
      name: templateId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ResourceId' }
...
    VersionId:
      name: versionId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ResourceId' }
...
  schemas:
...
    ReasonRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: { type: string, minLength: 2, maxLength: 500 }
...
    PolicyTemplateSummary:
      type: object
      additionalProperties: false
      required: [templateId, code, name, status, latestVersionNumber, rowVersion]
      properties:
        templateId: { $ref: '#/components/schemas/ResourceId' }
        code: { type: string }
        name: { type: string }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
        latestVersionNumber: { type: integer, minimum: 0 }
        rowVersion: { type: integer, format: int64, minimum: 0 }
    PolicyTemplateDetail:
      allOf:
        - $ref: '#/components/schemas/PolicyTemplateSummary'
        - type: object
          required: [description, fieldDefinitions]
          properties:
            description: { type: string }
            fieldDefinitions:
              type: array
              items:
                $ref: '#/components/schemas/PolicyFieldDefinition'
    PolicyTemplatePage:
      type: object
      additionalProperties: false
      required: [items, total, page, size]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PolicyTemplateSummary'
        total: { type: integer, format: int64, minimum: 0 }
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 100 }
    PolicyTemplateCreateRequest:
      type: object
      additionalProperties: false
      required: [code, name, description, fieldDefinitions]
      properties:
        code: { type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$' }
        name: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, maxLength: 500 }
        fieldDefinitions:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PolicyFieldDefinition'
...
    PolicyVersionDetail:
      allOf:
        - $ref: '#/components/schemas/PolicyVersionSummary'
        - type: object
          required: [parameters, scopeBindings, validation]
          properties:
            parameters:
              type: array
              items:
                $ref: '#/components/schemas/PolicyParameterValue'
            scopeBindings:
              type: array
              items:
                $ref: '#/components/schemas/PolicyScopeBinding'
            validation:
              $ref: '#/components/schemas/PolicyValidationResult'
            snapshotDigest: { type: [string, 'null'] }
            rollbackOfVersionId:
              oneOf:
                - $ref: '#/components/schemas/ResourceId'
                - type: 'null'
    PolicyVersionPage:
      type: object
      additionalProperties: false
      required: [items, total, page, size]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PolicyVersionSummary'
        total: { type: integer, format: int64, minimum: 0 }
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 100 }
    PolicyDraftCreateRequest:
      type: object
      additionalProperties: false
      required: [effectiveFrom, changeReason]
      properties:
        basedOnVersionId:
          oneOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
        effectiveFrom: { type: string, format: date }
        effectiveTo: { type: [string, 'null'], format: date }
        changeReason: { type: string, minLength: 2, maxLength: 500 }
    PolicyDraftUpdateRequest:
      type: object
      additionalProperties: false
      required: [parameters, effectiveFrom, changeReason, expectedVersion]
      properties:
        parameters:
          type: array
          items:
            $ref: '#/components/schemas/PolicyParameterValue'
        effectiveFrom: { type: string, format: date }
        effectiveTo: { type: [string, 'null'], format: date }
        changeReason: { type: string, minLength: 2, maxLength: 500 }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
...
    ScopeBindingsUpdateRequest:
      type: object
      additionalProperties: false
      required: [bindings, expectedVersion]
      properties:
        bindings:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ScopeBindingRequest'
        expectedVersion: { type: integer, format: int64, minimum: 0 }
...
    PolicyValidationResult:
      type: object
      additionalProperties: false
      required: [valid, issues]
      properties:
        valid: { type: boolean }
        issues:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
        validatedAt: { type: [string, 'null'], format: date-time }
...
    PolicyConflictResult:
      type: object
      additionalProperties: false
      required: [hasConflicts, conflicts]
      properties:
        hasConflicts: { type: boolean }
        conflicts:
          type: array
          items:
            $ref: '#/components/schemas/PolicyConflict'
    PolicySimulationRequest:
      type: object
      additionalProperties: false
      required: [sampleName, inputs]
      properties:
        sampleName: { type: string, minLength: 1, maxLength: 100 }
        inputs:
          type: object
          additionalProperties: true
    PolicySimulationResult:
      type: object
      additionalProperties: false
      required: [matched, resolvedParameters, explanation]
      properties:
        matched: { type: boolean }
        resolvedParameters:
          type: array
          items:
            $ref: '#/components/schemas/PolicyParameterValue'
        explanation:
          type: array
          items: { type: string }
    PolicyImpactPreview:
      type: object
      additionalProperties: false
      required: [scopeCount, affectedObjectCount, effectiveFrom, frozenPeriodProtected, warnings]
      properties:
        scopeCount: { type: integer, minimum: 0 }
        affectedObjectCount: { type: integer, minimum: 0 }
        effectiveFrom: { type: string, format: date }
        effectiveTo: { type: [string, 'null'], format: date }
        frozenPeriodProtected: { type: boolean }
        warnings:
          type: array
          items: { type: string }
    PublishPolicyRequest:
      type: object
      additionalProperties: false
      required: [reason, expectedVersion]
      properties:
        reason: { type: string, minLength: 2, maxLength: 500 }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
    RollbackPolicyRequest:
      type: object
      additionalProperties: false
      required: [targetVersionId, reason, expectedVersion]
      properties:
        targetVersionId: { $ref: '#/components/schemas/ResourceId' }
        reason: { type: string, minLength: 2, maxLength: 500 }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
...
  responses:
...
    AccessDenied:
      description: 已认证但动作、数据范围、记录状态或会话条件禁止
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ResourceNotAvailable:
      description: 对象不存在或主体无权读取，响应不区分原因
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ValidationFailed:
      description: 请求字段、受控枚举或业务校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: 状态、范围、生效期、冻结保护或唯一性冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    StaleVersion:
      description: 乐观锁版本过期
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...